alteracao da alteracao na resposta do delete

+++ trabalho_api/trabalho_api/src/main/java/com/trabalho/api/services/ProdutoService.java
@@ -64,8 +64,13 @@ public class ProdutoService {
         return Optional.empty();
     }
 
-    public Optional<Void> deletar(Long id) {
-        repository.deleteById(id);
-        return Optional.of(null);
+    public Boolean deletar(Long id) {
+
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+
+        return false;
     }
 }
